+++ src/backend/chat/serializers.py
@@ -1,6 +1,9 @@
 """Serializers for chat application."""
 
+from django.conf import settings
+
 from django_pydantic_field.rest_framework import SchemaField  # pylint: disable=no-name-in-module
+from drf_spectacular.utils import extend_schema_field
 from rest_framework import serializers
 
 from chat import models
@@ -57,6 +60,13 @@ class ChatConversationRequestSerializer(serializers.Serializer):
         default=False,
         help_text="Force web search.",
     )
+    model_hrid = serializers.CharField(
+        required=False,
+        default=None,
+        help_text="HRID of the model to use for the conversation.",
+        allow_blank=True,
+        trim_whitespace=True,
+    )
 
     def update(self, instance, validated_data):
         """Update method is not applicable in this context."""
@@ -71,3 +81,42 @@ class ChatConversationRequestSerializer(serializers.Serializer):
         if value not in ["text", "data"]:
             raise serializers.ValidationError("Protocol must be either 'text' or 'data'.")
         return value
+
+    def validate_model_hrid(self, value):
+        """Validate the model_hrid field."""
+        value = value or None  # Convert empty string to None
+
+        if value and value not in settings.LLM_CONFIGURATIONS:
+            raise serializers.ValidationError("Invalid model_hrid.")
+
+        return value
+
+
+class LLModelSerializer(serializers.Serializer):  # pylint: disable=abstract-method
+    """Serializer for LL models."""
+
+    hrid = serializers.CharField(help_text="Human-readable ID of the model.")
+    model_name = serializers.CharField(help_text="Name of the model.")
+    human_readable_name = serializers.CharField(help_text="Human-readable name of the model.")
+    icon = serializers.CharField(
+        help_text="Icon representing the model.",
+        allow_blank=True,
+        required=False,
+    )
+
+    # Computed field to indicate if the model is the default model
+    is_default = serializers.SerializerMethodField(
+        help_text="Indicates if the model is the default model.",
+    )
+
+    @staticmethod
+    @extend_schema_field(serializers.BooleanField)
+    def get_is_default(obj) -> bool:
+        """Check if the model is the default model."""
+        return obj.hrid == settings.LLM_DEFAULT_MODEL_HRID
+
+
+class LLMConfigurationSerializer(serializers.Serializer):  # pylint: disable=abstract-method
+    """Serializer for LLM configuration."""
+
+    models = LLModelSerializer(many=True)