+++ src/backend/core/factories.py
@@ -22,6 +22,7 @@ class UserFactory(factory.django.DjangoModelFactory):
         # Skip postgeneration save, no save is made in the postgeneration methods.
         skip_postgeneration_save = True
 
+    allow_conversation_analytics = factory.Faker("boolean", chance_of_getting_true=50)
     sub = factory.Sequence(lambda n: f"user{n!s}")
     email = factory.Faker("email")
     full_name = factory.Faker("name")